catalog-service/requirements.txt
# FastAPI framework and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# MongoDB driver
motor==3.4.0
pymongo==4.6.1

# Data validation and serialization
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1
python-dateutil==2.8.2

# Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.8

# Networking
aiohttp==3.9.1
aiodns==3.1.1
charset-normalizer==3.3.2

# Performance optimizations
orjson==3.9.10
